requirements.txt
openai>=1.30.0
llm-guard>=0.3.14
guardrails-ai>=0.5.0
requests>=2.31.0
networkx>=3.2
gitpython>=3.1.40
tree-sitter>=0.21.0
python-dotenv>=1.0.0
rich>=13.7.0
packaging>=23.0
toml>=0.10.2
